pyproject.toml
[project]
name = "ux570-importer"
version = "1.0.0"
description = "Import recordings from Sony ICD-UX570 Digital Voice Recorder with date-based organization"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "PyQt6>=6.4",
]

[project.scripts]
dvr-import = "dvr_import:main"
dvr-import-gui = "dvr_gui:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
